+++ game/controllers/free_actor_controller.cpp
@@ -3,7 +3,6 @@
 //all rights reserved uwu
 
 #include "free_actor_controller.h"
-#include "game.h"
 #include "input.h"
 
 using namespace Verse;