+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Fire component
 - Procedural fire animation
 - White and Perlin Noise generator
+- Free camera
+#### Changed
+- Camera component and system
 
 ### **[0.1.3]** ~ Serialization _2021-05-07_
 #### Added